# api/openapi/notif.yaml
openapi: 3.0.0
servers:
  - url: //api.dev.tribe.tymetribe.com
    description: Default Server
info:
  version: 1.0.0
  title: Tribe API Documentation
  termsOfService: "http://swagger.io/terms/"
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"

paths:
  /notif/health:
    get:
      operationId: GetNotifHealthInformation
      summary: Check Notification service status
      description: Use for liveliness and readiness in k8s cluster
      responses:
        "200":
          description: OK
        "404":
          description: Service cannot find the requested resource
        "503":
          description: Service is not ready to handle the request
      tags:
        - notif
      security: []
  /notif/ws:
    post:
      operationId: SubscribeNotification
      summary: Subscribe to Notification Web (Type representation only)
      description: Describe the notification types in Websocket channel
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/SocialMessageCreated"
                  - $ref: "#/components/schemas/Ping"
                  - $ref: "#/components/schemas/EndorsementRequestCreated"
                  - $ref: "#/components/schemas/EducationCourseProgressRemind"
        "404":
          description: Service cannot find the requested resource
        "503":
          description: Service is not ready to handle the request
      tags:
        - notif
      security:
        - JWT: []
  /notif/notifications:
    get:
      operationId: GetListUserNotifications
      summary: Get Current User's Notifications
      description: Use for liveliness and readiness in k8s cluster
      parameters:
        - name: token
          in: query
          description: the next_token is returned in previous query to get the next page
          required: true
          example: 6f3d61e4-5fee-4b5a-995a-02f65765fb32
          schema:
            type: string
        - name: limit
          in: query
          description: the limitation of records will be returned
          required: true
          example: 10
          schema:
            type: integer
            maximum: 100
            minimum: 5
            format: int
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Error"
                  - type: object
                    properties:
                      notifications:
                        type: array
                        items:
                          $ref: "#/components/schemas/Notification"
                      next_token:
                        type: string
                        example: 6f3d61e4-5fee-4b5a-995a-02f65765fb32
                        description: the token will be use to get next page
        "404":
          description: Service cannot find the requested resource
        "503":
          description: Service is not ready to handle the request
      tags:
        - notif
      security:
        - JWT: []

components:
  securitySchemes:
    JWT:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
      description: tribe authentication token
  schemas:
    Event:
      type: string
      enum:
        - social.message.created
        - endorsement.request.created
        - education.course.progress.remind
    SocialMessageCreated:
      required:
        - sender
        - msg
        - event
        - channel_id
        - create_at
      properties:
        event:
          $ref: "#/components/schemas/Event"
        sender:
          type: object
          required:
            - id
            - first_name
            - last_name
            - image_url
          properties:
            id:
              description: id of message sender
              type: string
              example: 6f3d61e4-5fee-4b5a-995a-02f65765fb32
            first_name:
              type: string
              example: Damwon
              description: user first name
            last_name:
              type: string
              example: Kia
              description: user last name
            image_url:
              type: string
              example: ./example.a.png
              description: avatar of sender
        msg:
          type: string
          example: hello from DK
          description: message content
        channel_id:
          description: the channel id that have message
          type: string
          example: IKE_NGU
        create_at:
          description: time when user send message
          type: string
          example: "2023-05-25T10:50:50.764Z"
    Ping:
      type: string
      example: ping
    EndorsementRequestCreated:
      required:
        - endorsement_id
        - sender
        - event
        - create_at
      properties:
        endorsement_id:
          type: integer
          description: id of endorsement
          example: 10
        sender:
          type: object
          required:
            - id
            - first_name
            - last_name
            - image_url
          properties:
            id:
              description: id of endorsement sender
              type: string
              example: 6f3d61e4-5fee-4b5a-995a-02f65765fb32
            first_name:
              type: string
              example: Damwon
              description: user first name
            last_name:
              type: string
              example: Kia
              description: user last name
            image_url:
              type: string
              example: ./example.a.png
              description: avatar of sender
        event:
          $ref: "#/components/schemas/Event"
        create_at:
          description: time when user send message
          type: string
          example: "2023-05-25T10:50:50.764Z"
    Notification:
      properties:
        id:
          type: string
          description: the id of notification
          example: 6f3d61e4-5fee-4b5a-995a-02f65765fb32
        data:
          type: object
          example: example
          description: the data of notification. It depends on the event type
        is_seen:
          type: boolean
          example: false
          description: describe use is seen the notification or not
        event:
          type: string
          description: the event id
          example: social.message.created
        sender_id:
          type: string
          description: the id of the sender
          example: 6f3d61e4-5fee-4b5a-995a-02f65765fb32
    Error:
      properties:
        slug:
          type: string
          example: user-not-found
      required:
        - slug
    EducationCourseProgressRemind:
      required:
        - source
        - event
        - create_at
      properties:
        event:
          $ref: "#/components/schemas/Event"
        course:
          description: information about the course
          type: object
          properties:
            id:
              type: integer
              description: id of the course
              example: 10
            title:
              type: string
              description: title of the course
              example: facebook tutorial
          required:
            - id
            - title
        create_at:
          description: time when user send message
          type: string
          example: "2022-12-22T06:00:00.200Z"

tags:
  - name: notif
    description: Notification Service

x-tagGroups:
  - name: Notification Service
    tags:
      - notif