# internal/auth/openapi_types.gen.go
// Package main provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.12.4 DO NOT EDIT.
package main

// ChangePasswordRequestBody defines model for ChangePasswordRequestBody.
type ChangePasswordRequestBody struct {
	// AccessToken cognito access token
	AccessToken string `json:"access_token"`

	// NewPassword old password
	NewPassword string `json:"new_password"`

	// OldPassword old password
	OldPassword string `json:"old_password"`
}

// ConfirmSignUpRequestBody defines model for ConfirmSignUpRequestBody.
type ConfirmSignUpRequestBody struct {
	// Code verification code
	Code string `json:"code"`

	// Email email that user used to sign up
	Email string `json:"email"`
}

// Error defines model for Error.
type Error struct {
	Slug string `json:"slug"`
}

// ForgotPasswordRequestBody defines model for ForgotPasswordRequestBody.
type ForgotPasswordRequestBody struct {
	// Email user email
	Email string `json:"email"`
}

// ForgotPasswordResponse defines model for ForgotPasswordResponse.
type ForgotPasswordResponse struct {
	// DeliveryMedium device that will receive code
	DeliveryMedium *string `json:"delivery_medium,omitempty"`

	// Destination email address will receive the code
	Destination *string `json:"destination,omitempty"`
}

// LoginRequestBody defines model for LoginRequestBody.
type LoginRequestBody struct {
	// Email user email
	Email string `json:"email"`

	// Password user password
	Password string `json:"password"`
}

// LoginResponse defines model for LoginResponse.
type LoginResponse struct {
	AccessToken  string `json:"access_token"`
	ExpireIn     int64  `json:"expire_in"`
	IdToken      string `json:"id_token"`
	RefreshToken string `json:"refresh_token"`
	TokenType    string `json:"token_type"`
}

// RefreshTokenRequestBody defines model for RefreshTokenRequestBody.
type RefreshTokenRequestBody struct {
	// RefreshToken refresh token to get new tokens
	RefreshToken string `json:"refresh_token"`
}

// RefreshTokenResponse defines model for RefreshTokenResponse.
type RefreshTokenResponse struct {
	AccessToken string `json:"access_token"`
	ExpireIn    int64  `json:"expire_in"`
	IdToken     string `json:"id_token"`
	TokenType   string `json:"token_type"`
}

// ResendCodeRequestBody defines model for ResendCodeRequestBody.
type ResendCodeRequestBody struct {
	// Email email address will receive the code
	Email string `json:"email"`
}

// ResendCodeResponse defines model for ResendCodeResponse.
type ResendCodeResponse struct {
	// DeliveryMedium device that will receive code
	DeliveryMedium *string `json:"delivery_medium,omitempty"`

	// Destination email address will receive the code
	Destination *string `json:"destination,omitempty"`
}

// SignUpRequestBody defines model for SignUpRequestBody.
type SignUpRequestBody struct {
	// Email user email
	Email string `json:"email"`

	// FirstName user first name
	FirstName string `json:"first_name"`

	// LastName user last name
	LastName string `json:"last_name"`

	// Password user password
	Password string `json:"password"`
}

// SignUpResponse defines model for SignUpResponse.
type SignUpResponse struct {
	Destination string `json:"destination"`
	UserSub     string `json:"user_sub"`
}

// ResendCodeJSONRequestBody defines body for ResendCode for application/json ContentType.
type ResendCodeJSONRequestBody = ResendCodeRequestBody

// ChangePasswordJSONRequestBody defines body for ChangePassword for application/json ContentType.
type ChangePasswordJSONRequestBody = ChangePasswordRequestBody

// ConfirmForgotPasswordJSONRequestBody defines body for ConfirmForgotPassword for application/json ContentType.
type ConfirmForgotPasswordJSONRequestBody = ForgotPasswordRequestBody

// ForgotPasswordJSONRequestBody defines body for ForgotPassword for application/json ContentType.
type ForgotPasswordJSONRequestBody = ForgotPasswordRequestBody

// RefreshTokenJSONRequestBody defines body for RefreshToken for application/json ContentType.
type RefreshTokenJSONRequestBody = RefreshTokenRequestBody

// ConfirmSignUpJSONRequestBody defines body for ConfirmSignUp for application/json ContentType.
type ConfirmSignUpJSONRequestBody = ConfirmSignUpRequestBody

// LoginJSONRequestBody defines body for Login for application/json ContentType.
type LoginJSONRequestBody = LoginRequestBody

// SignUpJSONRequestBody defines body for SignUp for application/json ContentType.
type SignUpJSONRequestBody = SignUpRequestBody